вынес список в отдельный компонент переименовал pages в app и добавил стилей
