+++ src/app/index.ts
@@ -1,0 +1,2 @@
+export * from './taskList';
+export * from './taskForm';
