+++ src/app/taskList/list/List.types.ts
@@ -1,0 +1,5 @@
+import type { TaskLocal } from '../../../types/task/TaskLocal.types';
+export type ListProps = {
+  tasks: TaskLocal[];
+  navigate: (value: string) => void;
+};
