+++ src/app/taskList/task/Task.module.css
@@ -1,0 +1,11 @@
+.task {
+  width: 100%;
+  margin-bottom: 5%;
+  border: 1px solid black;
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+  border-radius: 5px;
+  padding: 0 2%;
+  box-shadow: 0 0 10px 2px rgba(0, 0, 0, 0.3);
+}
