+++ src/app/taskList/task/Task.tsx
@@ -1,0 +1,6 @@
+import { TaskProps } from './Task.types';
+import style from './Task.module.css';
+
+export function Task({ children }: TaskProps) {
+  return <div className={style.task}>{children}</div>;
+}
