--- src/pages/index.ts
@@ -1,2 +1,0 @@
-export * from './taskList/TaskList';
-export * from './taskForm/TaskForm';
