--- src/pages/taskList/task/Task.tsx
@@ -1,5 +1,0 @@
-import { TaskProps } from './Task.types';
-
-export function Task({ children }: TaskProps) {
-  return <div className="task">{children}</div>;
-}
